Add a specific setting for flash IO mode.
The previous implementation of a generic set of flash tool arguments
wasn't ideal; for esptool.py, for example, the order of the arguments
is important, with some arguments being required in specific poisitions
and disallowed in others. Therefore, I switched the implementation for
a specific argument for flash mode (which is unspecified if left at the
default).